pipeline/requirements.txt
# AWS SDK and Services
boto3>=1.34.0          # AWS SDK (Bedrock, S3)
botocore>=1.34.0       # AWS core library

# HTTP and Web
requests>=2.31.0       # HTTP client for SEC EDGAR
httpx>=0.25.0          # Modern HTTP client (alternative)

# Data Processing
pydantic>=2.5.0        # Data validation
python-dotenv>=1.0.0   # Environment variables

# PDF Processing
PyPDF2>=3.0.0          # PDF parsing
pdfplumber>=0.10.0     # Alternative PDF parser

# Database
psycopg2-binary>=2.9.9 # PostgreSQL adapter
sqlalchemy>=2.0.23     # ORM (optional, for easier queries)

# HTML/XML Parsing
beautifulsoup4>=4.12.0 # HTML parsing (for EDGAR)
lxml>=5.1.0            # XML parsing

# Utilities
retry>=0.9.2           # Retry decorator
tenacity>=8.2.3        # Advanced retry logic
python-dateutil>=2.8.2 # Date parsing

# Testing (optional)
pytest>=7.4.3          # Testing framework
pytest-asyncio>=0.21.0 # Async test support
